spark/requirements.txt
pyspark==4.0.0
pytest
pytest-cov
awsglue3-local
